Rename composer class and refactor tests.
