+++ src/ServiceProvider.php
@@ -19,7 +19,7 @@ class ServiceProvider extends BaseServiceProvider
 {
     public function register(): void
     {
-        $this->app->singleton(Composer::class);
+        $this->app->singleton(ComposerBag::class);
         $this->app->singleton(ResponseFactory::class);
         $this->app->bind(Gateway::class, HttpGateway::class);
